+++ Source/Podio .NET/Models/SearchResult.cs
@@ -7,9 +7,9 @@ namespace PodioAPI.Models
 {
     public class SearchResult
     {
-       
+
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonProperty("type")]
         public string Type { get; set; }
@@ -43,5 +43,5 @@ namespace PodioAPI.Models
         public JToken Highlight { get; set; }
     }
 
-   
+
 }
